added option to skip Konsave

+++ 2-user.sh
@@ -67,11 +67,17 @@ for PKG in "${PKGS[@]}"; do
     yay -S --noconfirm $PKG
 done
 
-yay -S --noconfirm konsave
-pip install konsave
-konsave -i $HOME/ArchMO/atmoz.knsv
-sleep 1
-konsave -a atmoz
+read -p "Do you want to restore KONSAVE? [y/n]" answer
+if [[ $answer = y ]] ; then
+    yay -S --noconfirm konsave
+    pip install konsave
+    konsave -i $HOME/ArchMO/atmoz.knsv
+    sleep 1
+    konsave -a atmoz
+fi
+if [[ $answer = n ]] ; then
+    echo "Skipping KONSAVE"
+fi
 
 echo -e "\nDone!\n"
 exit
